refactor(tests): separate asset I/O integration
